+++ src/backend/x11/mod.rs
@@ -8,4 +8,5 @@
 pub mod compositor;
 #[path = "compositor/common/mod.rs"]
 pub mod compositor_common;
+pub(crate) mod scheduling;
 pub mod wm;